refactor(examples): use importable url_shortener package name (#94)
Rename examples/url-shortener to examples/url_shortener so tests can
import the example directly (matching examples/todo). Remove the
in-example test shim and simplify tests/unit/url_shortener/.
