+++ examples/url_shortener/domain/dtos/link_dto.py
