+++ examples/url_shortener/domain/protocols/link_repository_protocol.py
