+++ examples/url_shortener/domain/services/link_service.py
