+++ examples/url_shortener/infrastructure/database/models/link_model.py
