+++ examples/url_shortener/infrastructure/repositories/link_repository.py
