+++ examples/url_shortener/interface/server/schemas/link_schema.py
